+++ examples/01_empty_window/main.py
@@ -23,7 +23,7 @@ frame_count = 0
 fps_timer = 0.0
 
 while window.is_open():
-    window.poll_events()
+    bz.poll_events()
 
     ctx.begin_frame()